requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
transformers
sentence-transformers
scikit-learn
numpy<2
streamlit
plotly
pandas
typing-extensions